refactor(server): :recycle: better mybatis db request support

+++ src/main/java/com/heyrudy/app/core/interactors/DeleteTodoByIdUsecase.java
@@ -1,6 +1,7 @@
 package com.heyrudy.app.core.interactors;
 
 import com.heyrudy.app.core.port.actions.ITodoRepository;
+import com.heyrudy.app.exception.DbRequestException;
 import org.springframework.stereotype.Component;
 
 @Component
@@ -13,6 +14,8 @@ public final class DeleteTodoByIdUsecase {
     }
 
     public void execute(final Long id) {
+        if(repository.getTodoById(id).isEmpty())
+            throw new DbRequestException(String.format("todo with id : %d " + "don't exist in Database", id));
         repository.deleteTodoById(id);
     }
 }
